# docs/api/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "description": "S3C_TOKEN 中的 Bearer token。可通过 S3C_TOKEN_SCOPES 为单个 token 限定 readonly / prefixes（桶或桶内键前缀）/ accounts / expiresAt；越权请求返回 403，过期 token 返回 401。未在 S3C_TOKEN_SCOPES 中登记的 token 为全权。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "s3client API",
    "version": "v1.0.0",
    "description": "s3client 是面向 S3 兼容对象存储的多账号 Web 控制台。本文档为 /api/* 端点的 OpenAPI 3.0 契约，所有响应均 JSON（除 /api/health 等纯状态端点）。鉴权：Bearer Token（环境变量 S3C_TOKEN，多值逗号分隔）。"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/migrate": {
      "post": {
        "operationId": "migrate",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "sourceAccountId": "1f0c2a44-0b1e-4f5a-9c3d-7e8f9a0b1c2d",
                "sourceBucket": "src-bucket",
                "sourceKeys": [
                  "docs/a.txt"
                ],
                "targetAccountId": "2a1b3c4d-5e6f-7081-92a3-b4c5d6e7f809",
                "targetBucket": "dst-bucket",
                "targetPrefix": "migrated/"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "sourceAccountId": {
                    "type": "string"
                  },
                  "sourceBucket": {
                    "type": "string"
                  },
                  "sourceKeys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "targetAccountId": {
                    "type": "string"
                  },
                  "targetBucket": {
                    "type": "string"
                  },
                  "targetPrefix": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "migrated": 1,
                  "failed": 0,
                  "failedKeys": []
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "failed": {
                      "type": "integer"
                    },
                    "failedKeys": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "lastError": {
                      "type": "string"
                    },
                    "migrated": {
                      "type": "integer"
                    }
                  }
                }
              }
            },
            "description": "含 migrated/failed/failedKeys/lastError"
          }
        },
        "summary": "同步迁移（流式）",
        "tags": [
          "migrate"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/",
      "description": "同源（前端 Vite 代理或后端 SPA fallback）"
    }
  ],
...
}